+++ Jaberah/Models/ViewModels/Reports/GetBestStudentsReportForView.cs
@@ -6,8 +6,8 @@
         public string? GroupName { get; set; }
         public double SaveGrade { get; set; }
         public double ReviewGrade { get; set; }
-        public int AttendanceGrade { get; set; }
-        public int BehaviorGrade { get; set; }
+        public double AttendanceGrade { get; set; }
+        public double BehaviorGrade { get; set; }
         public double OralGrade { get; set; }
         public double PaperGrade { get; set; }
         public double Total { get; set; }
