+++ Jaberah/Models/ViewModels/Reports/GetSemesterReportForView.cs
@@ -5,8 +5,8 @@
         public int StudentId { get; set; }
         public string StudentName { get; set; } = string.Empty;
         public double GradeSum { get; set; }
-        public int AttendanceSum { get; set; }
-        public int BehaviorSum { get; set; }
+        public double AttendanceSum { get; set; }
+        public double BehaviorSum { get; set; }
         public double OralGradeSum { get; set; }
         public double PaperGradeSum { get; set; }
         public double MidFinalGrade { get; set; }
